Update APIs and enhance tool borrow transaction handling
Refactor API endpoints for login and registration, add transaction handling logic for tool borrowing, and improve UI features such as tool condition display, pull-to-refresh, and Snackbar feedback. Enhance the HomeScreen and ToolDetails to handle tool statuses consistently.

note that the swipe to refresh functionality is not working due to z-positioning of the components (BorrowRequestTool)